+++ conan/tools/microsoft/__init__.py
@@ -1,0 +1,3 @@
+# noinspection PyUnresolvedReferences
+from .msbuild import MSBuild
+from .msbuildtoolchain import MSBuildToolchain
